Rewrite the usecases to actually working scenarios
Only one of the usecases was clearly marked as not working in this form,
while in reality all but the first do not work.
This could have led to confusion for prospective users of the module or
worse turn them away from using the module.

As the current puppet recipes are proven via CI to actually support these
usecases on Ubuntu Precise and RedHat 6.4 family, instead of adding a
header about them not being expected to work, they were instead adapted
to working examples (with minor adaptions to user specific setups as a
caveat).

Change-Id: Id30d339ef1295bb50dcf017eefc996edfc8dbe60
Closes-Bug: #1327785